+++ src/github.ts
@@ -17,3 +17,9 @@ export type Context = {
 }
 
 export const getContext = (): Context => github.context
+export type Issue = {
+  owner: string
+  repo: string
+  number: number
+  body?: string // lazy
+}
