+++ src/run.ts
@@ -1,7 +1,6 @@
 import * as core from '@actions/core'
 import { appendOrUpdateBody } from './body.js'
-import { Context, getOctokit, Octokit } from './github.js'
-import { Issue } from './types.js'
+import { Context, getOctokit, Issue, Octokit } from './github.js'
 import { RequestError } from '@octokit/request-error'
 
 export type Inputs = {
@@ -20,12 +19,15 @@ type Operations = {
 
 export const run = async (inputs: Inputs, context: Context): Promise<void> => {
   const octokit = getOctokit(inputs.token)
-  const { owner, repo } = context.repo
-  const issues = inputs.issueNumbers.map((number) => ({ owner, repo, number }))
+
+  const issues = inputs.issueNumbers.map((number) => ({
+    owner: context.repo.owner,
+    repo: context.repo.repo,
+    number,
+  }))
 
   if (inputs.context) {
-    const pulls = await inferPullRequestFromContext(octokit, context)
-    issues.push(...pulls)
+    issues.push(...(await inferIssuesFromContext(octokit, context)))
   }
 
   for (const issue of issues) {
@@ -39,9 +41,9 @@ export const run = async (inputs: Inputs, context: Context): Promise<void> => {
   }
 }
 
-const inferPullRequestFromContext = async (octokit: Octokit, context: Context): Promise<Issue[]> => {
+const inferIssuesFromContext = async (octokit: Octokit, context: Context): Promise<Issue[]> => {
   if (Number.isSafeInteger(context.issue.number)) {
-    core.info(`inferred #${context.issue.number} from the current context`)
+    core.info(`Inferred #${context.issue.number} from the current workflow run`)
     return [
       {
         owner: context.repo.owner,
@@ -51,7 +53,7 @@ const inferPullRequestFromContext = async (octokit: Octokit, context: Context): 
     ]
   }
 
-  core.info(`list pull request(s) associated with ${context.sha}`)
+  core.info(`List pull request(s) associated with ${context.sha}`)
   const pulls = await octokit.rest.repos.listPullRequestsAssociatedWithCommit({
     owner: context.repo.owner,
     repo: context.repo.repo,
@@ -63,7 +65,7 @@ const inferPullRequestFromContext = async (octokit: Octokit, context: Context): 
     number: pr.number,
     body: pr.body || '',
   }))
-  core.info(`inferred pull requests: ${issues.map((i) => `#${i.number}`).join(`, `)}`)
+  core.info(`Using pull requests: ${issues.map((i) => `#${i.number}`).join(`, `)}`)
   return issues
 }
 
@@ -75,21 +77,25 @@ const processIssue = async (octokit: Octokit, r: Operations, issue: Issue): Prom
       issue_number: issue.number,
       labels: r.addLabels,
     })
-    core.info(`added label(s) to #${issue.number}: ${added.map((label) => label.name).join(', ')}`)
+    core.info(
+      `Added label(s) to ${issue.owner}/${issue.repo}#${issue.number}: ${added.map((label) => label.name).join(', ')}`,
+    )
   }
 
   for (const labelName of r.removeLabels) {
     try {
-      const { data: removed } = await octokit.rest.issues.removeLabel({
+      await octokit.rest.issues.removeLabel({
         owner: issue.owner,
         repo: issue.repo,
         issue_number: issue.number,
         name: labelName,
       })
-      core.info(`removed label ${labelName} from #${issue.number}: ${removed.map((label) => label.name).join(', ')}`)
+      core.info(`Removed label ${labelName} from ${issue.owner}/${issue.repo}#${issue.number}`)
     } catch (error) {
       if (error instanceof RequestError && error.status === 404) {
-        core.warning(`could not remove label ${labelName} from #${issue.number}: ${error.message}`)
+        core.warning(
+          `Could not remove label ${labelName} from ${issue.owner}/${issue.repo}#${issue.number}: ${error.message}`,
+        )
         continue
       }
       throw error
@@ -103,7 +109,7 @@ const processIssue = async (octokit: Octokit, r: Operations, issue: Issue): Prom
       issue_number: issue.number,
       body: r.postComment,
     })
-    core.info(`create a comment to #${issue.number}: ${created.html_url}`)
+    core.info(`Created a comment to ${issue.owner}/${issue.repo}#${issue.number}: ${created.html_url}`)
   }
 
   if (r.appendOrUpdateBody) {
