--- src/types.ts
@@ -1,6 +1,0 @@
-export type Issue = {
-  owner: string
-  repo: string
-  number: number
-  body?: string // lazy
-}
